+++ src/employees/employee.controller.ts
@@ -13,7 +13,7 @@ import {
 } from '@nestjs/common';
 import { CreateEmployeeDto } from './dto/create-employee.dto';
 import { UpdateEmployeeDto } from './dto/update-employee.dto';
-import { Employee } from './entities/employees.entity';
+import { Employee } from './entities/employee.entity';
 import { EmployeeService } from './employee.service';
 import { JwtAuthGuard } from 'src/auth/auth.guard';
