+++ src/employees/employee.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { EmployeeService } from './employee.service';
 import { EmployeeController } from './employee.controller';
-import { Employee } from './entities/Employees.entity';
+import { Employee } from './entities/employee.entity';
 import { Employee_Saving } from './entities/employee_saving.entity';
 import { EmployeeRepository } from './employee.repository';
