+++ src/employees/employee.repository.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@nestjs/common';
 import { DataSource, Repository } from 'typeorm';
 import { CreateEmployeeDto } from './dto/create-employee.dto';
 import { UpdateEmployeeDto } from './dto/update-employee.dto';
-import { Employee } from './entities/Employees.entity';
+import { Employee } from './entities/employee.entity';
 
 @Injectable()
 export class EmployeeRepository {
